fix bug of xtuner

+++ xtuner/v1/rl/base/worker.py
@@ -326,7 +326,7 @@ class TrainingWorker(SingleAcceleratorWorker):
                     f"rollout_logprobs {rollout_logprobs.shape} vs old_logprobs {old_logprobs.shape}"
                 )
                 if rollout_logprobs.numel() == 0:  # pad 情况下是空的
-                    min_diff = torch.tensor(0)
+                    min_diff = torch.tensor(0).float()
                     max_diff = min_diff
                     std_diff = min_diff
                     mean_diff = min_diff
@@ -335,7 +335,7 @@ class TrainingWorker(SingleAcceleratorWorker):
                     max_diff = torch.max(rollout_logprobs - old_logprobs)
                     mean_diff = torch.mean(rollout_logprobs - old_logprobs)
                     if rollout_logprobs.numel() == 1:
-                        std_diff = torch.tensor(0)
+                        std_diff = torch.tensor(0).float()
                     else:
                         std_diff = torch.std(rollout_logprobs - old_logprobs)
                 all_diffs.append((min_diff, max_diff, mean_diff, std_diff))
